Fix: ParentNode couldn't add new nodes to branches with args
This fix let me refactor the plugin handling to enable them
when the appropriate `set <plugin>` is called, whether in
.ideavimrc or by typing it. Unfortunately, the nodes don't yet
provide a way to de-register, and I'm not quite familiar enough
with them yet to implement it, so `set no<plugin>` does not work.

`ys` commands now actually work in the editor!