Fix quote support (misread the plugin); improve repeat slightly
Motions like `t;` did in fact use the args of args stuff, so I
wrapped the arg in a subclass on which I could attach extra
data relevant to the action. Seems like a good approach that could
support things like surround-with-tag. Should probably do something
similar for the `cs`