First stab at <Plug> mappings
Using a non-existent key stroke is great, but ambiguous
plug mappings aren't supported out of the box. Imagine you
have two mappings, both of which take a motion:

<Plug>More
<Plug>MoreMagic

You've mapped gmm to <Plug>More and gmc to <Plug>MoreMagic.

When you type gmc, the recursive mapping is available, and the
`else if (mappingInfo != null)` branch can feed the mapped keys
back into the system. Because these are fed as recursive, when
the keys come back through, they can get the mapping to MoreMagic
(which is no longer a prefix) and execute.

For gmm, when it feeds the keys for <Plug>More, it is still
a prefix; when a motion is inputted, it is no longer a prefix,
and so it would hit the `else` block, feeding the keys WITHOUT
recursion. So, we have to catch the case where we're no longer
a prefix, but can find a <Plug> mapping within the input. This
seems to work pretty well, but causes "recursive records found"
warnings in the console, and complaints about local history being
broken. I'm not entirely sure whether this is just an idiosyncrasy
of the unit test environment, or whether this would cause
significant problems in the live environment.

Also included is some prep work for <Plug>-based repeat handling.
Extra input sequences past the <Plug> mapping are stored in a
PlugInputModel which is read from in preference to the
ModalInputDialog.