Fix VIM-636 Broken behavior with shift-s to substitute line
|S| aka |cc| was not behaving properly on files that were empty or
contained 1 to 3 blank lines.

I have found the problem for the 2 and 3 line files to be not counting
newlines when getting of the file size in deleteLine. I have given it
some thought and we are comparing to an offset which includes newlines.
So I am pretty confident that using getFileSize(editor, true) is more
accurate regardless of the fix.

For the empty and one blank line problems I have added special case
checks. I did it this way because I was able to confine the change only
for change line thus minimizing the breakage risk.