VIM-714 Fixed down movement column after visual delete.
It seems that the movements that were done while in visual selection
mode were also moving the editor column information.

Despite the above it seems that original vim will use the column of the
position where the deletion is started, so this is what  I have
implemented to fix this.