Fix skipping leading whitespace
`ap` motion selects paragraphs differently depending on whether the caret is placed on whitespace or paragraph text. ie when it is placed on text it will select the text and the whitespace after the paragraph text. When placed on whitespace it will select the next paragraph thus the whitespace is placed before the paragraph instead of after.

Previously the code assumed that whitespace will always be after the text and thus missing the leading whitespace.