Moved Andrey Vlasovskikh to previous maintainers
I maintained the plugin during 2012-10/2021-02. In 2019-02 Alex Plate joined me as a plugin developer and maintainer. Since then Alex has added tons of cool features and important fixes and has become the #1 committer to the IdeaVim repository. Thank you Alex for your great work! I'm stepping down from my maintainer role, but I will keep contributing to the plugin. Alex Plate is now the maintainer of IdeaVim. There is a live community of contributors to the plugin mentioned in AUTHORS.md. IdeaVim is a very active and welcoming open-source project. I wish Alex and the IdeaVim community all the best!